Create ETL_Python
This is a Python code where it takes a XML file, parses it and extract the information needed, then uploads it to a AWS based SQL database. The XML files are from NIH PubMed website. There are 1184 XML files in total and each of them contains 30000 publication information, including abstract, pmid, title etc